otesbro/requirements.txt
absl-py==0.11.0
argon2-cffi==20.1.0
asgiref==3.3.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.4
djangorestframework==3.12.2
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gast==0.3.3
gdown==3.11.1
gluoncv==0.7.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.34.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.3.0
ipykernel==5.4.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.2
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mxnet==1.6.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
ninja==1.10.0.post2
notebook==6.1.6
numpy==1.18.2
oauthlib==3.1.0
opencv-contrib-python==4.2.0.34
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.8
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
portalocker==2.0.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.3
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
pywebview==3.2
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.15.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.1
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
torch==1.5.0
torchvision==0.2.2.post3
tornado==6.1
tqdm==4.43.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.0